requirements.txt
Django==2.2.5
django-bootstrap-daterangepicker==1.0.4
pkg-resources==0.0.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
